Fixing the block bindings enqueue
